WA-271 Extracted structs to files, reordered vars and funcs
